Fixes #5 and possibly other issues
Issue noticed due to multiple aliased belongsTo relationships to the same external model. The problem was with the `_decryptArray()` method and how it referred to the current class name, which unfortunately in a recursive situation seems to not hold the expected value. By passing the object to the method we can more soundly rely on the value we are checking against.